małe poprawki

+++ pr-dl-cli.py
@@ -1,11 +1,15 @@
 #!/usr/bin/env python
 #-*- coding: utf-8 -*-
 import os
+import os.path
 import json
 import re
 import sys
 import pycurl
 import urllib2
+
+from os.path import expanduser
+home = expanduser("~")
 
 class PobierzStrone:
     def __init__(self):
@@ -30,20 +34,27 @@ if len(sys.argv) > 1:
             c.perform()
             c.close()
             # 2.1. Szukamy w stary sposób
-            # onclick="javascript:playFile('1083019','ee169737-54bb-40e0-b7be-eb4a69b7d4b8','','1057','/80/1007/Artykul/1141701,Musi-sie-udac-Joanna-Boguslawska','0','0');
-            # http://static.polskieradio.pl/ee169737-54bb-40e0-b7be-eb4a69b7d4b8.mp3            
             links = [] + re.findall('onclick="javascript:playFile\(\'[^\']*\',\'([^\']*)\',\'[^\']*\',\'[^\']*\',\'[^\']*\',\'1\',\'[^\']*\'\);', www.contents)
             titles = [] + re.findall('onclick="javascript:playFile\(\'[^\']*\',\'[^\']*\',\'([^\']*)\',\'[^\']*\',\'[^\']*\',\'1\',\'[^\']*\'\);', www.contents)
-            filenames = [] + re.findall('onclick="javascript:playFile\(\'[^\']*\',\'[^\']*\',\'[^\']*\',\'[^\']*\',\'([^\']*)\',\'1\',\'[^\']*\'\);', www.contents)
+            # print www.contents
             # 2.2. Szukamy w nowy sposób
-            # iSpeaker" data-media='{"id":874754,"file":"http://static.polskieradio.pl/4a9904eb-246e-4d7c-829d-32ac6ad05f9a.mp3","provider":"audio","uid":"4a9904eb-246e-4d7c-829d-32ac6ad05f9a","length":1800,"autostart":true,"link":"/39/0/Artykul/863703,Ronald-Reagan-czyli-jak-aktor-filmowy-zostal-prezydentem","title":"Ronald%20Reagan%2C%20czyli%20jak%20aktor%20filmowy%20zosta%C5%82%20prezydentem"
-            links = links + re.findall('iSpeaker" data-media=\'{"id":[0-9]*,"file":"[^"]*","provider":"audio","uid":"([^"]*)","length":[0-9]*,"autostart":true,"link":"[^"]*","title":"[^"]*"', www.contents)
-            titles = titles + re.findall('iSpeaker" data-media=\'{"id":[0-9]*,"file":"[^"]*","provider":"audio","uid":"[^"]*","length":[0-9]*,"autostart":true,"link":"[^"]*","title":"([^"]*)"', www.contents)
-            filenames = filenames + re.findall('iSpeaker" data-media=\'{"id":[0-9]*,"file":"[^"]*","provider":"audio","uid":"[^"]*","length":[0-9]*,"autostart":true,"link":"([^"]*)","title":"[^"]*"', www.contents)
-            if(len(links) == len(titles) and len(links) == len(filenames) and len(links) > 0):
+            # data-media={"id":1084883,"file":"http://static.polskieradio.pl/45b8b22e-2f2e-49c2-a99b-48fef8e4fa85.mp3","provider":"audio","uid":"45b8b22e-2f2e-49c2-a99b-48fef8e4fa85","length":873,"autostart":true,"link":"/39/0/Artykul/1145828,Kongres-Wiedenski-–-narodziny-nowoczesnej-dyplomacji","title":"Kongres%20Wiede%C5%84ski%20%E2%80%93%20narodziny%20nowoczesnej%20dyplomacji"
+            links = links + re.findall('data-media={"id":[0-9],"file":"([^"]*)","provider":"audio","uid":"[^"]*","length":[0-9]*,"autostart":[a-z]*,"link":"[^"]*","title":"[^"]*"', www.contents)
+            titles = titles + re.findall('data-media={"id":[0-9],"file":"[^"]*","provider":"audio","uid":"[^"]*","length":[0-9]*,"autostart":[a-z]*,"link":"[^"]*","title":"([^"]*)"', www.contents)
+            if(len(links)==0):
+                links = re.findall('"file":"([^"]*)"',www.contents)
+                # print len(links)
+                # print links
+                titles = re.findall('"title":"([^"]*)"',www.contents)
+                # print len(titles)
+                # print titles            
+            if(len(links) == len(titles) and len(links) > 0):
                 a = 0
                 while a < len(links):
-                    url = 'http://static.polskieradio.pl/'+links[a]+'.mp3'
+                    url = links[a]
+                    target_dir = home+'/'
+                    if not url.find('static.polskieradio.pl'):
+                        url = 'http://static.polskieradio.pl/'+url+'.mp3'
                     print (str(a) + ". " + url)
                     title = titles[a]
                     title = urllib2.unquote(title)
@@ -57,26 +68,27 @@ if len(sys.argv) > 1:
                     title = title.replace('_-_','-')
                     print 'Tytuł: '+title
                     if(title != ''):
-                        file_name = title+'.mp3'
+                        file_name = target_dir + title+'.mp3'
                     else:
-                        file_name = links[a]+'.mp3'
+                        file_name = target_dir + links[a]+'.mp3'
                     u = urllib2.urlopen(url)
-                    f = open(file_name, 'wb')
-                    meta = u.info()
-                    file_size = int(meta.getheaders("Content-Length")[0])
-                    print("Pobieranie: {0} Rozmiar: {1}".format(url, file_size))            
-                    file_size_dl = 0
-                    block_sz = 8192
-                    while True:
-                        buffer = u.read(block_sz)
-                        if not buffer:
-                            break
-                        file_size_dl += len(buffer)
-                        f.write(buffer)
-                        p = float(file_size_dl) / file_size
-                        status = r"{0}  [{1:.2%}]".format(file_size_dl, p)
-                        status = status + chr(8)*(len(status)+1)                
-                        sys.stdout.write(status)
-                    f.close()
+                    if not os.path.isfile(file_name):
+                        f = open(file_name, 'wb')
+                        meta = u.info()
+                        file_size = int(meta.getheaders("Content-Length")[0])
+                        print("Pobieranie: {0} Rozmiar: {1}".format(url, file_size))            
+                        file_size_dl = 0
+                        block_sz = 8192
+                        while True:
+                            buffer = u.read(block_sz)
+                            if not buffer:
+                                break
+                            file_size_dl += len(buffer)
+                            f.write(buffer)
+                            p = float(file_size_dl) / file_size
+                            status = r"{0}  [{1:.2%}]".format(file_size_dl, p)
+                            status = status + chr(8)*(len(status)+1)                
+                            sys.stdout.write(status)
+                        f.close()
                     a += 1
